rabbitMQ_api/consumer/requirements.txt
pika==1.1.0
Flask==1.1.1
#numpy==1.16.2
#pandas==0.24.2
#pylint==2.3.1
python-dotenv==0.12.0